+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog of threedi-modelchecker
 
 - Add check that warns when the total percentage mapped to the same surface via SurfaceMap is greater than 100% (nens/rana#3863)
 - Add check that warns when the sum of the weights in MeasureMap for a single control are not 1+/-0.01 (nens/rana#3597)
+- Add check that warns when a closed culvert or pipe is embedded (nens/threedi-modelchecker#408)
 
 
 2.18.18 (2026-04-10)